Adding interactive part for manual testing
There is at least one serious problem here. I think one key issue is that it is analyzing words in general rather than the letters in the words. Another issue is a lack of training data. Yet another issue is that there are so many words across the different languages that a dictionary size of 10000 may not be enough to adequately represent them.