Increasing learning rate and epoch count, changing to softmax output
The softmax has the benefit of being interpretable as a probability density function. The increased learning rate and number of epochs seem to drastically improve the performance of the model, to the point where it is now comparable in performance to the naive version I made years ago based on the letter frequency of words.